app/src/main/java/com/github/user/capstonesoilit/utils/DateUtils.kt
package com.github.user.capstonesoilit.utils

import java.text.SimpleDateFormat
import java.util.*

object DateUtils {
    fun formatDateTime(dateTime: String): String {
        val inputFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", Locale.getDefault())
        val outputFormat = SimpleDateFormat("HH/dd/MM/yy", Locale.getDefault())

        return try {
            val date = inputFormat.parse(dateTime)
            outputFormat.format(date!!)
        } catch (e: Exception) {
            dateTime
        }
    }
}
